Переделала MainEmpty из страницы в компонент
