--- project/src/pages/main-empty-screen/main-empty-screen.tsx
@@ -1,30 +1,0 @@
-import Header from '../../components/header/header';
-import Filter from '../../components/filter/filter';
-import { useAppSelector } from '../../hooks';
-import { getSelectedCity } from '../../store/offer-process/selectors';
-
-const MainEmptyScreen = (): JSX.Element => {
-  const city = useAppSelector(getSelectedCity);
-  return (
-    <div className='page page--gray page--main'>
-      <Header />
-      <main className='page__main page__main--index page__main--index-empty'>
-        <h1 className='visually-hidden'>Cities</h1>
-        <Filter />
-        <div className='cities'>
-          <div className='cities__places-container cities__places-container--empty container'>
-            <section className='cities__no-places'>
-              <div className='cities__status-wrapper tabs__content'>
-                <b className='cities__status'>No places to stay available</b>
-                <p className='cities__status-description'>We could not find any property available at the moment in {city}</p>
-              </div>
-            </section>
-            <div className='cities__right-section'></div>
-          </div>
-        </div>
-      </main>
-    </div>
-  );
-};
-
-export default MainEmptyScreen;
